requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3
pydantic==2.6.1